Rename profiling anr package to trigger
